Avoid useless middleware file
